requirements.txt
requests==2.26.0
tabulate==0.8.9
colorama==0.4.4
python-bidi==0.4.2
bs4==0.0.1




